+++ pybamm/models/submodels/interface/kinetics/base_kinetics.py
@@ -79,6 +79,9 @@ class BaseKinetics(BaseInterface):
             ]
         else:
             ocp = variables[f"{Domain} electrode {reaction_name}open circuit potential"]
+        # If ocp was broadcast, take only the orphan.
+        if isinstance(ocp, pybamm.Broadcast):
+            ocp = ocp.orphans[0]
         eta_r = delta_phi - ocp
 
         # Get average interfacial current density