+++ pybamm/models/submodels/interface/lithium_plating/plating.py
@@ -22,7 +22,10 @@ class Plating(BasePlating):
     .. [1] SEJ O'Kane, ID Campbell, MWJ Marzook, GJ Offer and M Marinescu. "Physical
            Origin of the Differential Voltage Minimum Associated with Li Plating in
            Lithium-Ion Batteries". Journal of The Electrochemical Society,
-           167:090540, 2019
+           167:090540, 2020
+    .. [2] SEJ O'Kane, W Ai, G Madabattula, D Alonso-Alvarez, R Timms, V Sulzer,
+           JS Edge, B Wu, GJ Offer and M Marinescu. "Lithium-ion battery degradation:
+           how to model it". Physical Chemistry: Chemical Physics, 24:7909, 2022
 
     **Extends:** :class:`pybamm.lithium_plating.BasePlating`
     """
@@ -31,6 +34,7 @@ class Plating(BasePlating):
         super().__init__(param, options)
         self.x_average = x_average
         pybamm.citations.register("OKane2020")
+        pybamm.citations.register("OKane2022")
 
     def get_fundamental_variables(self):
         if self.x_average is True:
@@ -39,14 +43,24 @@ class Plating(BasePlating):
                 domain="current collector",
             )
             c_plated_Li = pybamm.PrimaryBroadcast(c_plated_Li_av, "negative electrode")
+            c_dead_Li_av = pybamm.Variable(
+                "X-averaged dead lithium concentration",
+                domain="current collector",
+            )
+            c_dead_Li = pybamm.PrimaryBroadcast(c_dead_Li_av, "negative electrode")
         else:
             c_plated_Li = pybamm.Variable(
                 "Lithium plating concentration",
                 domain="negative electrode",
                 auxiliary_domains={"secondary": "current collector"},
             )
+            c_dead_Li = pybamm.Variable(
+                "Dead lithium concentration",
+                domain="negative electrode",
+                auxiliary_domains={"secondary": "current collector"},
+            )
 
-        variables = self._get_standard_concentration_variables(c_plated_Li)
+        variables = self._get_standard_concentration_variables(c_plated_Li, c_dead_Li)
 
         return variables
 
@@ -59,20 +73,26 @@ class Plating(BasePlating):
         c_plated_Li = variables["Lithium plating concentration"]
         j0_stripping = param.j0_stripping(c_e_n, c_plated_Li, T)
         j0_plating = param.j0_plating(c_e_n, c_plated_Li, T)
+        # phi_ref is part of the de-dimensionalization used in PyBaMM
         phi_ref = param.n.prim.U_ref / param.potential_scale
 
         eta_stripping = delta_phi + phi_ref + eta_sei
         eta_plating = -eta_stripping
-        prefactor = 1 / (2 * (1 + self.param.Theta * T))
+        prefactor = 1 / (1 + self.param.Theta * T)
+        # NEW: transfer coefficients can be set by the user
+        alpha_stripping = self.param.alpha_stripping
+        alpha_plating = self.param.alpha_plating
 
-        if self.options["lithium plating"] == "reversible":
+        if self.options["lithium plating"] in ["reversible", "partially reversible"]:
             j_stripping = j0_stripping * pybamm.exp(
-                prefactor * eta_stripping
-            ) - j0_plating * pybamm.exp(prefactor * eta_plating)
+                prefactor * alpha_stripping * eta_stripping
+            ) - j0_plating * pybamm.exp(prefactor * alpha_plating * eta_plating)
         elif self.options["lithium plating"] == "irreversible":
             # j_stripping is always negative, because there is no stripping, only
             # plating
-            j_stripping = -j0_plating * pybamm.exp(prefactor * eta_plating)
+            j_stripping = -j0_plating * pybamm.exp(
+                prefactor * alpha_plating * eta_plating
+            )
 
         variables.update(self._get_standard_overpotential_variables(eta_stripping))
         variables.update(self._get_standard_reaction_variables(j_stripping))
@@ -85,26 +105,41 @@ class Plating(BasePlating):
     def set_rhs(self, variables):
         if self.x_average is True:
             c_plated_Li = variables["X-averaged lithium plating concentration"]
+            c_dead_Li = variables["X-averaged dead lithium concentration"]
             j_stripping = variables[
                 "X-averaged lithium plating interfacial current density"
             ]
-            # Note a is dimensionless (has a constant value of 1 if the surface
-            # area does not change)
             a = variables["X-averaged negative electrode surface area to volume ratio"]
+            L_sei = variables["X-averaged total SEI thickness"]
         else:
             c_plated_Li = variables["Lithium plating concentration"]
+            c_dead_Li = variables["Dead lithium concentration"]
             j_stripping = variables["Lithium plating interfacial current density"]
             a = variables["Negative electrode surface area to volume ratio"]
+            L_sei = variables["Total SEI thickness"]
 
         Gamma_plating = self.param.Gamma_plating
+        # In the partially reversible plating model, coupling term turns reversible
+        # lithium into dead lithium. In other plating models, it is zero.
+        if self.options["lithium plating"] == "partially reversible":
+            dead_lithium_decay_rate = self.param.dead_lithium_decay_rate(L_sei)
+            coupling_term = dead_lithium_decay_rate * c_plated_Li
+        else:
+            coupling_term = pybamm.Scalar(0)
 
-        self.rhs = {c_plated_Li: -Gamma_plating * a * j_stripping}
+        self.rhs = {
+            c_plated_Li: -Gamma_plating * a * j_stripping - coupling_term,
+            c_dead_Li: coupling_term,
+        }
 
     def set_initial_conditions(self, variables):
         if self.x_average is True:
             c_plated_Li = variables["X-averaged lithium plating concentration"]
+            c_dead_Li = variables["X-averaged dead lithium concentration"]
         else:
             c_plated_Li = variables["Lithium plating concentration"]
+            c_dead_Li = variables["Dead lithium concentration"]
         c_plated_Li_0 = self.param.c_plated_Li_0
+        zero = pybamm.Scalar(0)
 
-        self.initial_conditions = {c_plated_Li: c_plated_Li_0}
+        self.initial_conditions = {c_plated_Li: c_plated_Li_0, c_dead_Li: zero}