+++ pybamm/parameters/lithium_ion_parameters.py
@@ -157,9 +157,12 @@ class LithiumIonParameters(BaseParameters):
         self.k_sei_dim = pybamm.Parameter("SEI kinetic rate constant [m.s-1]")
         self.U_sei_dim = pybamm.Parameter("SEI open-circuit potential [V]")
 
-        # Li plating parameters
+        # Lithium plating parameters
         self.V_bar_plated_Li = pybamm.Parameter(
             "Lithium metal partial molar volume [m3.mol-1]"
+        )
+        self.c_Li_typ = pybamm.Parameter(
+            "Typical plated lithium concentration [mol.m-3]"
         )
         self.c_plated_Li_0_dim = pybamm.Parameter(
             "Initial plated lithium concentration [mol.m-3]"
@@ -219,6 +222,11 @@ class LithiumIonParameters(BaseParameters):
         return pybamm.FunctionParameter(
             "Exchange-current density for plating [A.m-2]", inputs
         )
+
+    def dead_lithium_decay_rate_dimensional(self, L_sei):
+        """Dimensional dead lithium decay rate [s-1]"""
+        inputs = {"Total SEI thickness [m]": L_sei}
+        return pybamm.FunctionParameter("Dead lithium decay rate [s-1]", inputs)
 
     def _set_scales(self):
         """Define the scales used in the non-dimensionalisation scheme"""
@@ -389,10 +397,10 @@ class LithiumIonParameters(BaseParameters):
         self.c_sei_init = self.c_ec_0_dim / self.c_sei_outer_scale
 
         # lithium plating parameters
-        self.c_Li_typ = pybamm.Parameter(
-            "Typical plated lithium concentration [mol.m-3]"
-        )
         self.c_plated_Li_0 = self.c_plated_Li_0_dim / self.c_Li_typ
+
+        self.alpha_plating = pybamm.Parameter("Lithium plating transfer coefficient")
+        self.alpha_stripping = 1 - self.alpha_plating
 
         # ratio of lithium plating reaction scaled to intercalation reaction
         self.Gamma_plating = (
@@ -466,6 +474,12 @@ class LithiumIonParameters(BaseParameters):
         return (
             self.j0_plating_dimensional(c_e_dim, c_Li_dim, T_dim) / self.n.prim.j_scale
         )
+
+    def dead_lithium_decay_rate(self, L_sei):
+        """Dimensionless exchange-current density for stripping"""
+        L_sei_dim = L_sei * self.L_sei_0_dim
+
+        return self.dead_lithium_decay_rate_dimensional(L_sei_dim) * self.timescale
 
     def _set_input_current(self):
         """Set the input current"""
@@ -887,6 +901,11 @@ class ParticleLithiumIonParameters(BaseParameters):
 
     def U_dimensional(self, sto, T, lithiation=None):
         """Dimensional open-circuit potential [V]"""
+        # bound stoichiometry between tol and 1-tol. Adding 1/sto + 1/(sto-1) later
+        # will ensure that ocp goes to +- infinity if sto goes into that region
+        # anyway
+        tol = 1e-10
+        sto = pybamm.maximum(pybamm.minimum(sto, 1 - tol), tol)
         if lithiation is None:
             lithiation = ""
         else: